// src/app/shared/models/burrito.model.ts
export class Burrito {
    key: string;
    restaurant_name: string;
    order: string;
    cost: number;
    comments: string;
    meat_type: string;
    meat: number;
    beans_type: string;
    beans: number;
    rice: number;
    guacamole_type: string;
    guacamole: number;
    crema: number;
    salsa: number;
    tortilla: number;
    cheese: number;
    other: string;
    size: number;
    taste: number;
    overall_rating: number;
    picture: string;
    date: Date;
    yelp_link: string;
}